[DDW-367] use empty theme as default to avoid duplicated styles
