+++ __tests__/Bubble.test.js
@@ -1,11 +1,11 @@
 import React from 'react';
-import renderer from 'react-test-renderer';
 
 import { Bubble } from '../source/components/Bubble';
 import { BubbleSkin } from '../source/skins/simple/BubbleSkin';
+import { renderInSimpleTheme } from './helpers/theming';
 
 test('Bubble renders correctly', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <Bubble skin={BubbleSkin} />
   );
 
@@ -14,7 +14,7 @@ test('Bubble renders correctly', () => {
 });
 
 test('Bubble renders isOpeningUpward={true}', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <Bubble isOpeningUpward skin={BubbleSkin} />
   );
 
@@ -23,7 +23,7 @@ test('Bubble renders isOpeningUpward={true}', () => {
 });
 
 test('Bubble renders isTransparent={false}', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <Bubble isTransparent={false} skin={BubbleSkin} />
   );
 
@@ -32,7 +32,7 @@ test('Bubble renders isTransparent={false}', () => {
 });
 
 test('Bubble renders isHidden={true}', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <Bubble isHidden skin={BubbleSkin} />
   );
 
@@ -41,7 +41,7 @@ test('Bubble renders isHidden={true}', () => {
 });
 
 test('Bubble renders isFloating={true}', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <Bubble isFloating skin={BubbleSkin} />
   );
