+++ __tests__/Button.test.js
@@ -1,11 +1,11 @@
 import React from 'react';
-import renderer from 'react-test-renderer';
 
 import { Button } from '../source/components/Button';
 import { ButtonSkin } from '../source/skins/simple/ButtonSkin';
+import { renderInSimpleTheme } from './helpers/theming';
 
 test('Button renders correctly', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <Button skin={ButtonSkin} />
   );
 
@@ -14,7 +14,7 @@ test('Button renders correctly', () => {
 });
 
 test('Button renders with a label', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <Button label="send" skin={ButtonSkin} />
   );
 
@@ -23,7 +23,7 @@ test('Button renders with a label', () => {
 });
 
 test('Button is disabled', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <Button disabled skin={ButtonSkin} />
   );
