+++ __tests__/Checkbox.test.js
@@ -1,11 +1,11 @@
 import React from 'react';
-import renderer from 'react-test-renderer';
 
 import { Checkbox } from '../source/components/Checkbox';
 import { CheckboxSkin } from '../source/skins/simple/CheckboxSkin';
+import { renderInSimpleTheme } from './helpers/theming';
 
 test('Checkbox renders correctly', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <Checkbox skin={CheckboxSkin} />
   );
 
@@ -14,7 +14,7 @@ test('Checkbox renders correctly', () => {
 });
 
 test('Checkbox renders with a label', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <Checkbox label="check here" skin={CheckboxSkin} />
   );
 
@@ -23,7 +23,7 @@ test('Checkbox renders with a label', () => {
 });
 
 test('Checkbox is disabled', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <Checkbox disabled skin={CheckboxSkin} />
   );
 
@@ -32,7 +32,7 @@ test('Checkbox is disabled', () => {
 });
 
 test('Checkbox is checked', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <Checkbox checked skin={CheckboxSkin} />
   );
