+++ __tests__/FormField.test.js
@@ -1,13 +1,13 @@
 import React from 'react';
-import renderer from 'react-test-renderer';
 
 import { FormField } from '../source/components/FormField';
 import { FormFieldSkin } from '../source/skins/simple/FormFieldSkin';
+import { renderInSimpleTheme } from './helpers/theming';
 
 const renderFormField = () => <div className="render-prop" />;
 
 test('FormField renders correctly', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <FormField
       skin={FormFieldSkin}
       render={renderFormField}
@@ -19,7 +19,7 @@ test('FormField renders correctly', () => {
 });
 
 test('FormField renders with label', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <FormField
       label="Add a Label"
       skin={FormFieldSkin}
@@ -32,7 +32,7 @@ test('FormField renders with label', () => {
 });
 
 test('FormField renders with an error', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <FormField
       error="Add an Error"
       skin={FormFieldSkin}
@@ -45,7 +45,7 @@ test('FormField renders with an error', () => {
 });
 
 test('FormField is disabled', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <FormField
       disabled
       skin={FormFieldSkin}
@@ -58,7 +58,7 @@ test('FormField is disabled', () => {
 });
 
 test('FormField renders an input element', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <FormField
       skin={FormFieldSkin}
       render={() => <input className="render-prop" />}
