+++ __tests__/Input.test.js
@@ -1,11 +1,11 @@
 import React from 'react';
-import renderer from 'react-test-renderer';
 
 import { Input } from '../source/components/Input';
 import { InputSkin } from '../source/skins/simple/InputSkin';
+import { renderInSimpleTheme } from './helpers/theming';
 
 test('Input renders correctly', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <Input skin={InputSkin} />
   );
 
@@ -14,7 +14,7 @@ test('Input renders correctly', () => {
 });
 
 test('Input renders with placeholder', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <Input placeholder="0.0000" skin={InputSkin} />
   );
 
@@ -23,7 +23,7 @@ test('Input renders with placeholder', () => {
 });
 
 test('Input renders with a value', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <Input value="there is value" skin={InputSkin} />
   );
 
@@ -32,7 +32,7 @@ test('Input renders with a value', () => {
 });
 
 test('Input is readOnly', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <Input
       readOnly
       skin={InputSkin}
