+++ __tests__/Select.test.js
@@ -3,6 +3,7 @@ import renderer from 'react-test-renderer';
 
 import { Select } from '../source/components/Select';
 import { SelectSkin } from '../source/skins/simple/SelectSkin';
+import { renderInSimpleTheme } from './helpers/theming';
 
 const COUNTRIES = [
   { label: 'Frankreich', value: 'France' },
@@ -20,7 +21,7 @@ const COUNTRIES_DISABLED_OPTIONS = [
 ];
 
 test('Select renders correctly', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <Select
       options={COUNTRIES}
       skin={SelectSkin}
@@ -32,7 +33,7 @@ test('Select renders correctly', () => {
 });
 
 test('Select renders with placeholder', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <Select
       placeholder="Select your country …"
       options={COUNTRIES}
@@ -45,7 +46,7 @@ test('Select renders with placeholder', () => {
 });
 
 test('Select renders with an error', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <Select
       error="Please select a different option"
       options={COUNTRIES}
@@ -58,7 +59,7 @@ test('Select renders with an error', () => {
 });
 
 test('Select renders with disabled options', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <Select
       options={COUNTRIES_DISABLED_OPTIONS}
       skin={SelectSkin}
@@ -70,7 +71,7 @@ test('Select renders with disabled options', () => {
 });
 
 test('Select isOpeningUpward={true}', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <Select
       isOpeningUpward
       options={COUNTRIES}
@@ -83,7 +84,7 @@ test('Select isOpeningUpward={true}', () => {
 });
 
 test('Select isOpen={true}', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <Select
       isOpen
       options={COUNTRIES}
@@ -96,7 +97,7 @@ test('Select isOpen={true}', () => {
 });
 
 test('Select uses render prop - optionRenderer', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <Select
       options={COUNTRIES}
       skin={SelectSkin}
