+++ __tests__/TextArea.test.js
@@ -3,9 +3,10 @@ import renderer from 'react-test-renderer';
 
 import { TextArea } from '../source/components/TextArea';
 import { TextAreaSkin } from '../source/skins/simple/TextAreaSkin';
+import { renderInSimpleTheme } from './helpers/theming';
 
 test('TextArea renders correctly', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <TextArea skin={TextAreaSkin} />
   );
 
@@ -14,7 +15,7 @@ test('TextArea renders correctly', () => {
 });
 
 test('TextArea renders with placeholder', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <TextArea
       placeholder="0.0000"
       skin={TextAreaSkin}
@@ -26,7 +27,7 @@ test('TextArea renders with placeholder', () => {
 });
 
 test('TextArea renders with an error', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <TextArea
       error="Please enter valid input"
       skin={TextAreaSkin}
@@ -38,7 +39,7 @@ test('TextArea renders with an error', () => {
 });
 
 test('TextArea renders with a value', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <TextArea
       value="this one has a value"
       skin={TextAreaSkin}
@@ -50,7 +51,7 @@ test('TextArea renders with a value', () => {
 });
 
 test('TextArea renders with 5 rows', () => {
-  const component = renderer.create(
+  const component = renderInSimpleTheme(
     <TextArea
       rows={5}
       skin={TextAreaSkin}
